requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==3.1.0
PyYAML==6.0.2
requests==2.31.0
